Tests: move pushd contextmanager to argparse_testlib and reuse it
